Signup popover owns url (#1576)
* move login url back into signup popover

* simplify zindex
