+++ frontends/mit-learn/src/page-components/SignupPopover/SignupPopover.tsx
@@ -1,9 +1,8 @@
 import React from "react"
+import { Popover, Typography, styled, ButtonLink } from "ol-components"
 import type { PopoverProps } from "ol-components"
-import styled from "@emotion/styled"
-import { Popover } from "../Popover/Popover"
-import Typography from "@mui/material/Typography"
-import { ButtonLink } from "../Button/Button"
+import * as urls from "@/common/urls"
+import { useLocation } from "react-router"
 
 const StyledPopover = styled(Popover)({
   width: "300px",
@@ -26,13 +25,11 @@ const Footer = styled.div({
 type SignupPopoverProps = Pick<
   PopoverProps,
   "anchorEl" | "onClose" | "placement"
-> & {
-  signupUrl?: string
-}
+>
 const SignupPopover: React.FC<SignupPopoverProps> = (props) => {
-  const { signupUrl, ...popoverProps } = props
+  const loc = useLocation()
   return (
-    <StyledPopover {...popoverProps} open={!!popoverProps.anchorEl}>
+    <StyledPopover {...props} open={!!props.anchorEl}>
       <HeaderText variant="subtitle2">
         Join {APP_SETTINGS.SITE_NAME} for free.
       </HeaderText>
@@ -41,7 +38,14 @@ const SignupPopover: React.FC<SignupPopoverProps> = (props) => {
         and follow your areas of interest.
       </BodyText>
       <Footer>
-        <ButtonLink href={signupUrl}>Sign Up</ButtonLink>
+        <ButtonLink
+          href={urls.login({
+            pathname: loc.pathname,
+            search: loc.search,
+          })}
+        >
+          Sign Up
+        </ButtonLink>
       </Footer>
     </StyledPopover>
   )
