+++ frontends/ol-components/src/components/LearningResourceExpanded/LearningResourceExpanded.tsx
@@ -144,7 +144,6 @@ type LearningResourceExpandedProps = {
   imgConfig: EmbedlyConfig
   onAddToLearningPathClick?: LearningResourceCardProps["onAddToLearningPathClick"]
   onAddToUserListClick?: LearningResourceCardProps["onAddToUserListClick"]
-  signupUrl?: string
 }
 
 const ImageSection: React.FC<{
@@ -324,7 +323,6 @@ const LearningResourceExpanded: React.FC<LearningResourceExpandedProps> = ({
   imgConfig,
   onAddToLearningPathClick,
   onAddToUserListClick,
-  signupUrl,
 }) => {
   const [selectedRun, setSelectedRun] = useState(resource?.runs?.[0])
 
@@ -427,7 +425,6 @@ const LearningResourceExpanded: React.FC<LearningResourceExpandedProps> = ({
         user={user}
         onAddToLearningPathClick={onAddToLearningPathClick}
         onAddToUserListClick={onAddToUserListClick}
-        signupUrl={signupUrl}
       />
     </Container>
   )
