+++ frontends/ol-components/src/components/ThemeProvider/ThemeProvider.tsx
@@ -9,7 +9,6 @@ import * as typography from "./typography"
 import * as buttons from "./buttons"
 import * as chips from "./chips"
 import { colors } from "./colors"
-import { MUI_DRAWER_Z_INDEX } from "../../constants/styleOverrides"
 
 const shadow = {
   shadowOffsetX: 3,
@@ -81,13 +80,6 @@ const themeOptions: ThemeOptions = {
       styleOverrides: { paper: { borderRadius: "4px" } },
     },
     MuiChip: chips.chipComponent,
-    MuiDrawer: {
-      styleOverrides: {
-        paper: {
-          zIndex: MUI_DRAWER_Z_INDEX,
-        },
-      },
-    },
   },
 }
